001 code cleanup
